fix firebase errors

+++ lib/widgets/slidable_todo.dart
@@ -18,16 +18,36 @@ class SlidableToDo extends StatelessWidget {
           color: Colors.red,
           icon: Icons.delete,
           onTap: () {
-            todo.inTrash = true;
-            // ignore: deprecated_member_use
-            Scaffold.of(context).showSnackBar(
-              SnackBar(
-                backgroundColor: Colors.red[700],
-                duration: Duration(seconds: 2),
-                content: Text('Todo moved to recycle bin'),
-              ),
-            );
-            firestoreService.saveToDo(todo);
+            showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertDialog(
+                    title: Text('Attention!'),
+                    content: Text('Do you want to move todo to recycle bin?'),
+                    actions: [
+                      TextButton(
+                          onPressed: () {
+                            todo.inTrash = true;
+                            // ignore: deprecated_member_use
+                       /*     Scaffold.of(context).showSnackBar(
+                              SnackBar(
+                                backgroundColor: Colors.red[700],
+                                duration: Duration(seconds: 2),
+                                content: Text('Todo moved to recycle bin'),
+                              ),
+                            ); */
+                            firestoreService.saveToDo(todo);
+                            Navigator.of(context).pop();
+                          },
+                          child: Text('Yes')),
+                      TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                          },
+                          child: Text('No'))
+                    ],
+                  );
+                });
           },
         )
       ];
@@ -39,13 +59,13 @@ class SlidableToDo extends StatelessWidget {
         onTap: () {
           todo.inTrash = false;
           // ignore: deprecated_member_use
-          Scaffold.of(context).showSnackBar(
+        /*  Scaffold.of(context).showSnackBar(
             SnackBar(
               backgroundColor: Colors.green,
               duration: Duration(seconds: 2),
               content: Text('Todo was restored'),
             ),
-          );
+          ); */
           firestoreService.saveToDo(todo);
         },
       ),
@@ -54,15 +74,35 @@ class SlidableToDo extends StatelessWidget {
         color: Colors.red,
         icon: Icons.delete,
         onTap: () {
-          // ignore: deprecated_member_use
-          Scaffold.of(context).showSnackBar(
-            SnackBar(
-              backgroundColor: Colors.redAccent[700],
-              duration: Duration(seconds: 2),
-              content: Text('Todo was deleted'),
-            ),
-          );
-          firestoreService.removeToDo('${todo.toDoId}');
+          showDialog(
+              context: context,
+              builder: (context) {
+                return AlertDialog(
+                  title: Text('Attention!'),
+                  content: Text('Do you want to delete this todo?'),
+                  actions: [
+                    TextButton(
+                        onPressed: () {
+                          // ignore: deprecated_member_use
+                        /*  Scaffold.of(context).showSnackBar(
+                            SnackBar(
+                              backgroundColor: Colors.redAccent[700],
+                              duration: Duration(seconds: 2),
+                              content: Text('Todo was deleted'),
+                            ),
+                          ); */
+                          firestoreService.removeToDo('${todo.toDoId}');
+                          Navigator.of(context).pop();
+                        },
+                        child: Text('Yes')),
+                    TextButton(
+                        onPressed: () {
+                          Navigator.of(context).pop();
+                        },
+                        child: Text('No'))
+                  ],
+                );
+              });
         },
       )
     ];
